feat(client): Add support for "can_link_account" Desktop Channel message.
This message is called in the Desktop context before a "sign in" or
"sign up" to check with the Desktop browser whether another user has
previously logged in to the browser. If one has, Desktop presents
a warning to the user, giving her an opportunity to cancel the operation.
If the user cancels, we cancel the underlying "sign in" and "sign up"
operation with the auth server.